// backend/src/coin-histories/coin-histories.types.ts
export enum TransactionHistoryType {
  CREDIT = 1,
  DEBIT = 0,
}

export enum TransactionHistorySource {
  BOUGHT = 1,
  REFERRAL = 2,
  ADS_REWARD = 3,
  FACEBOOK = 4,
  TWITTER = 5,
  RATING = 6,
  BIRTHDAY = 7,
  TRANSFER_TO = 8,
  TRANSFER_FROM = 9,
  USED = 10,
  GAME_ATTENDANCE_DEDUCTION = 11,
  SIGNUP_BONUS = 12,
}